Refactor footprint TSX generation into per-element converters with stable output ordering
