+++ lib/generate-footprint-tsx/convert-plated-holes.ts
@@ -1,0 +1,23 @@
+import { su } from "@tscircuit/soup-util"
+import { mmStr } from "@tscircuit/mm"
+import type { FootprintElementConverter } from "./converter-types"
+import { formatPcbRotationAttr } from "./helpers"
+
+export const convertPlatedHoles: FootprintElementConverter = (circuitJson) => {
+  const platedHoles = su(circuitJson).pcb_plated_hole.list()
+  const elementStrings: string[] = []
+
+  for (const platedHole of platedHoles) {
+    if (platedHole.shape === "oval" || platedHole.shape === "pill") {
+      elementStrings.push(
+        `<platedhole  portHints={${JSON.stringify(platedHole.port_hints)}} pcbX="${mmStr(platedHole.x)}" pcbY="${mmStr(platedHole.y)}" outerHeight="${mmStr(platedHole.outer_height)}" outerWidth="${mmStr(platedHole.outer_width)}" holeHeight="${mmStr(platedHole.hole_height)}" holeWidth="${mmStr(platedHole.hole_width)}" height="${mmStr(platedHole.hole_height)}" shape="${platedHole.shape}"${formatPcbRotationAttr("ccw_rotation" in platedHole ? platedHole.ccw_rotation : undefined)} />`,
+      )
+    } else if (platedHole.shape === "circle") {
+      elementStrings.push(
+        `<platedhole  portHints={${JSON.stringify(platedHole.port_hints)}} pcbX="${mmStr(platedHole.x)}" pcbY="${mmStr(platedHole.y)}" outerDiameter="${mmStr(platedHole.outer_diameter)}" holeDiameter="${mmStr(platedHole.hole_diameter)}" shape="circle" />`,
+      )
+    }
+  }
+
+  return elementStrings
+}
