+++ lib/generate-footprint-tsx/converter-types.ts
@@ -1,0 +1,5 @@
+import type { AnyCircuitElement } from "circuit-json"
+
+export type FootprintElementConverter = (
+  circuitJson: AnyCircuitElement[],
+) => string[]
